Updates Mailcontroller auth Auth

+++ Backend/Controllers/MailController.cs
@@ -1,13 +1,7 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Backend.Models;
-using Backend.Services;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Backend.Controllers
 {
@@ -25,71 +19,24 @@ namespace Backend.Controllers
             this._userManager = userManager;
         }
 
-        // GET: api/Mail
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<Mail>>> GetMails()
-        {
-            return await this._context.Mail.ToListAsync();
-        }
-
         // GET: api/Mail/5
         [HttpGet("{id}")]
+        [Authorize]
         public async Task<ActionResult<Mail>> GetMail(long id)
         {
-            var mail = await this._context.Mail.FindAsync(id);
-
-            if (mail == null)
-            {
+            Mail? mail = await this._context.Mail.FindAsync(id);
+            if (mail is null)
                 return this.NotFound();
-            }
 
+            var self = await this._userManager.GetUserAsync(this.User);
+            if (self is null || self.Id != mail.OwnerMailBox?.OwnerId)
+                return this.Forbid();
             return mail;
-        }
-
-        // PUT: api/Mail/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutMail(long id, Mail mail)
-        {
-            if (id != mail.Id)
-            {
-                return this.BadRequest();
-            }
-
-            this._context.Entry(mail).State = EntityState.Modified;
-
-            try
-            {
-                await this._context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!this.MailExists(id))
-                {
-                    return this.NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return this.NoContent();
-        }
-
-        // POST: api/Mail
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<ActionResult<Mail>> PostMail(Mail mail)
-        {
-            this._context.Mail.Add(mail);
-            await this._context.SaveChangesAsync();
-
-            return this.CreatedAtAction("GetMail", new { id = mail.Id }, mail);
         }
 
         // DELETE: api/Mail/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> DeleteMail(long id)
         {
             var mail = await this._context.Mail.FindAsync(id);
@@ -97,16 +44,14 @@ namespace Backend.Controllers
             {
                 return this.NotFound();
             }
+            var self = await this._userManager.GetUserAsync(this.User);
+            if (self is null || self.Id != mail.OwnerMailBox?.OwnerId)
+                return this.Forbid();
 
             this._context.Mail.Remove(mail);
             await this._context.SaveChangesAsync();
 
             return this.NoContent();
-        }
-
-        private bool MailExists(long id)
-        {
-            return this._context.Mail.Any(e => e.Id == id);
         }
     }
 }
